feat: move reusable functions to new file
