+++ static/claim-all.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Claim all permit</title>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <link href="out/styles/claim-permits/index.css" rel="stylesheet" />
+</head>
+
+<body>
+  <div class="full-centered">
+    <div class="container">
+      <h1 class="audit-tag">Claim all permit</h1>
+      <button type="button" class="btn btn-primary mb-3" id="claimAll">
+        <span class="btn-loader" id="report-loader"></span>
+        <span>Claim all</span>
+      </button>
+    </div>
+  </div>
+  <ul class="notifications"></ul>
+  <script src="out/scripts/claim-permits/claim.js" type="application/javascript"></script>
+</body>
+
+</html>
