+++ static/styles/claim-permits/index.css
@@ -1,0 +1,9 @@
+@import url(../audit-report/audit.css);
+
+.full-centered {
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
